Rename registration form
