+++ packages/shopware-auth/src/component/RegistrationForm/RegistrationForm.context.js
