+++ packages/shopware-auth/src/component/RegistrationForm/index.js
